# Reintroduce unsafe in encoding::string::merge