@danburkert 2019-06-17
The benchmark numbers above were run on the https://github.com/danburkert/prost/tree/benchmarks branch.  I'm putting some TLC into the benchmark suite so I can properly review #187, which lead to finding this regression.