@danburkert 2020-01-12
@mzabaluev 

> This still leaves the String value in a potentially invalid state if any method on buf panics.
> A scope guard, like originally proposed in 55f644e for #194 but working on the &mut Vec<u8> to > avoid technical unsoundness, would be more bulletproof.

Great point.  I've pushed a new implementation with a drop guard.  It's simpler than the original, since it turned out the implemention should have been clearing the string/bytes values on merge all along.  This was revealed to be an issue in the conformance tests during the move from protobuf 3.7 to 3.11.  I'm happy to report this drop guard impl has the same performance as without the drop guard.